calculate distance to trendingPoi locally

+++ HudHud/Extensions/DateFormatter.swift
@@ -20,7 +20,8 @@ extension String {
     }
 }
 
-extension Int {
+extension Double {
+
     func getDistanceString() -> String {
         let distanceFormatter = MeasurementFormatter()
         distanceFormatter.unitOptions = .providedUnit
